requirements.txt
numpy
pandas
streamlit
plotly
matplotlib
seaborn
scikit-learn
ipython
apache-airflow

